# backend/swagger.yaml
openapi: 3.0.0
paths:
...
  '/users/{id}':
    get:
      operationId: UsersController_findById
      summary: ''
      description: Return all properties for one user OR throw an error
      parameters:
        - name: id
          required: true
          in: path
          description: Id of the user we want to find in the database.
          allowEmptyValue: false
          examples:
            a:
              summary: User id is 1
              value: '1'
            b:
              summary: User id is 42
              value: '42'
          schema:
            type: string
        - name: friend
          required: true
          in: query
          description: Get the user friend ?
          allowEmptyValue: true
          examples:
            a:
              summary: '?friend=true'
              value: 'true'
            b:
              summary: '?friend=false'
              value: 'false'
          schema:
            type: string
        - name: channel
          required: true
          in: query
          description: Get the user channel ?
          allowEmptyValue: true
          examples:
            a:
              summary: '?channel=true'
              value: 'true'
            b:
              summary: '?channel=false'
              value: 'false'
          schema:
            type: string
      responses:
        '200':
          description: 'Return an object containing the user info + ?{friend/channel}'
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - username
                  - AvatarUrl
                properties:
                  id:
                    type: number
                  username:
                    type: string
                  avatarUrl:
                    type: string
                  friends:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                        username:
                          type: string
                        avatarUrl:
                          type: string
                  channelsProfiles:
                    type: array
                    items:
                      type: object
                      properties:
                        channelId:
                          type: number
                        role:
                          type: string
                        channel:
                          type: object
                          items:
                            type: object
                            properties:
                              channelName:
                                type: string
      tags:
        - Users
...
  /channels:
    post:
      operationId: ChannelsController_create
      summary: ''
      description: ' Create a new Channel AND a new UserOnChannel (with the current user who made the request as the userId and the newly created chan as chanId). Upon success returns the new chan infos.'
      parameters: []
      requestBody:
        required: true
        description: The Description for the Post Body. Please look into the DTO. You will see the @ApiOptionalProperty used to define the Schema.
        content:
          application/json:
            schema:
              type: string
            examples:
              good:
                summary: Basic Body
                value:
                  mode: PRIVATE
                  channelName: string
              empty:
                summary: Empty Body
                description: Empty
                value: {}
              withId:
                summary: With Id
                description: Check the id is not pris en compte.
                value:
                  mode: PRIVATE
                  channelName: string
                  id: 1
              PubWithPassword:
                summary: Pub With Password
                description: Check the Password is not pris en compte.
                value:
                  mode: PUBLIC
                  channelName: PubWithPass
                  password: JeSuisUnPassword
              PrivWithPassword:
                summary: Priv With Password
                description: Check the Password is not pris en compte.
                value:
                  mode: PRIVATE
                  channelName: PrivWithPass
                  password: JeSuisUnPassword
              ProtWithPassword:
                summary: Prot With Password
                description: Check the Password is pris en compte.
                value:
                  mode: PROTECTED
                  channelName: ProtWithPass
                  password: JeSuisUnPassword
              ProtWithNoPassword:
                summary: Prot With NO Password
                description: Should raise an error.
                value:
                  mode: PROTECTED
                  channelName: ProtWithNoPass
              PubWithNOPassword:
                summary: Pub With no Password
                description: Check the Password is not pris en compte.
                value:
                  mode: PUBLIC
                  channelName: PubWithPass
              withWrongType:
                summary: With WrongType
                description: Wrong attribute pass
                value:
                  mode: PUBLIC
                  channelName: string
                  nieh: llalalala
              withEmptyDefinedName:
                summary: With EmptyDefinedName
                description: Attribute channelName defined but only ws
                value:
                  mode: PUBLIC
                  channelName: '                   '
              withTrimableName:
                summary: With TrimableName
                description: Attribute channelName with spaces between words (check to see if those spaces are not trimmed)
                value:
                  mode: PUBLIC
                  channelName: '    lalala je veux etre TRIM               '
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelEntity'
      tags:
        - Channels
      security:
        - TransandanceToken: []
    get:
      operationId: ChannelsController_suscribedChannels
      summary: ''
      description: Returns a list of the user`s suscribed channels.
      parameters: []
      responses:
        '200':
          description: ''
      tags:
        - Channels
      security:
        - TransandanceToken: []
  '/channels/join/{chanId}':
    post:
      operationId: ChannelsController_joinChan
      summary: ''
      description: Join a channel (create a new useronChannel record). Only permitted to non BANNED users.
      parameters:
        - name: chanId
          required: true
          in: path
          schema:
            type: number
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: string
      tags:
        - Channels
      security:
        - TransandanceToken: []
  '/channels/{id}':
    patch:
      operationId: ChannelsController_update
      summary: ''
      description: 'Update a specified channel: change its password and/or mode. Only CREATOR can do it.'
      parameters:
        - name: chanId
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        description: The Description for the Post Body. Please look into the DTO. You will see the @ApiOptionalProperty used to define the Schema.
        content:
          application/json:
            schema:
              type: string
            examples:
              onlyName:
                summary: Only name update
                value:
                  channelName: HIHIHIHI
              onlyPasswd:
                summary: Only passwd update
                value:
                  password: niiiiieh
              pubToProtWithPass:
                summary: Public chan to Protected with Password
                value:
                  mode: PROTECTED
                  password: niiiiieh
              pubToProtNOPass:
                summary: 'Public chan to Protected WITHOUT Password: should raise an error.'
                value:
                  mode: PROTECTED
              protToPubWithPass:
                summary: Protected to public/private with password (shouldnt let password in db).
                value:
                  mode: PUBLIC
                  password: niehpasspouet
              protToPub:
                summary: Protected to public/private.
                value:
                  mode: PUBLIC
              empty:
                summary: Empty test
                value: {}
              wrongType:
                summary: wrongType test
                value:
                  lala: adfs
                  suisworngtye: fdsfsd
                  hashdf: ''
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags:
        - Channels
      security:
        - TransandanceToken: []
    delete:
      operationId: ChannelsController_remove
      summary: ''
      description: 'Delete the specified chan. Upon success, returns the deleted chan. Only CREATOR can do it.'
      parameters:
        - name: id
          required: true
          in: path
          examples:
            good:
              summary: good
              value: '1'
            bad:
              summary: bad
              value: 43
            empty:
              summary: empty
              value: ''
            neg:
              summary: negative
              value: '-12'
            string:
              summary: string
              value: hahdf
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelEntity'
      tags:
        - Channels
      security:
        - TransandanceToken: []
...
info:
  title: TRANSCENDENCE
  description: Our 42 project ft_transcendence API description.
  version: v1
  contact: {}
...
servers: []
components:
  securitySchemes:
    TransandanceToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Enter access token for transandance
  schemas:
...
    ChannelEntity:
      type: object
      properties:
        mode:
          type: string
          enum:
            - PRIVATE
            - PUBLIC
            - PROTECTED
        id:
          type: number
        createdAt:
          format: date-time
          type: string
        updateAt:
          format: date-time
          type: string
        channelName:
          type: string
          minLength: 5
          maxLength: 30
      required:
        - mode
        - id
        - createdAt
        - updateAt
        - channelName
...